ASM-4248 Zone configuration is skipping server's WWPN. And zone is not added to the zoneset
Due to recent refactoring changes and moving from device to apply model, module is trying to open SSH Session on the Brocade switch for each resource
There is a limitation on the number of connections that we can open on the switch (default is 2).

Due to above limitation, only first two resources were getting invoked and rest were getting missed.

Added transport.close for all the create and destroy operations been used by ASM Manifest file, so that we don't reach the threshold level of connections.